requirements.txt
streamlit==0.89.0
numpy==1.20.3
pandas==1.3.2
matplotlib==3.4.2
seaborn==0.11.2
pillow==8.0.0
scikit-learn==0.24.2